+++ packages/expo-cli/src/commands/eas-build/utils/expoUpdates/index.ts
@@ -1,0 +1,2 @@
+export { configureUpdatesAndroidAsync, setUpdatesVersionsAndroidAsync } from './android';
+export { configureUpdatesIOSAsync, setUpdatesVersionsIOSAsync } from './ios';